Moved print functions into seperate file
